+++ ci/common/neovim.sh
@@ -7,7 +7,13 @@ NEOVIM_REPO=${NEOVIM_REPO:-neovim/neovim}
 NEOVIM_BRANCH=${NEOVIM_BRANCH:-master}
 
 clone_neovim() {
-  rm -rf ${NEOVIM_DIR}
-  git clone --branch ${NEOVIM_BRANCH} --depth 1 git://github.com/${NEOVIM_REPO} ${NEOVIM_DIR}
+  if is_ci_build || ! [ -d ${NEOVIM_DIR} ] ; then
+    rm -rf ${NEOVIM_DIR}
+    git clone --branch ${NEOVIM_BRANCH} --depth 1 git://github.com/${NEOVIM_REPO} ${NEOVIM_DIR}
+  else
+    git --git-dir=${NEOVIM_DIR}/.git rev-parse HEAD >/dev/null \
+      || exit 1
+  fi
+
   NEOVIM_COMMIT=$(git --git-dir=${NEOVIM_DIR}/.git rev-parse HEAD)
 }
